Revert "segundo sprint"
This reverts commit 5e1b8914d98bf73a65afedf25bc1ccb50686f54d.

+++ Galax/src/juego/Entidad.java
@@ -12,18 +12,18 @@ public abstract class Entidad {
 	protected final int width = 75;
 	protected final int height = 75;
 
-	protected double velocidad;
+	protected int velocidad;
 
 	protected Point pos;
 
-	protected Entidad(double velocidad, int x, int y) {
+	protected Entidad(int velocidad, int x, int y) {
 		this.pos = new Point(x, y);
 		this.velocidad = velocidad;
 
 		this.image = new Icon[4];
 	}
 
-	public double getVelocidad() {
+	public int getVelocidad() {
 		return velocidad;
 	}
